Travis CI: Build Dependencies with GCC 5
When building OpenCV with GCC 6 we get an error where something is trying to #include_next <stdlib.h>.  This is apparently a known problem (see here: https://gcc.gnu.org/ml/gcc-bugs/2016-03/msg00880.html).  As a workaround we build dependencies with GCC 5 and then switch to GCC 6 for our build assuming the appropriate environment variable is set.